Move DISTRO_VERSION declaration to KAS repository
As the KAS repository is responsible for the overall versioning, the
DISTRO_VERSION should be defined there as well. Unfortunately, it is not
possible to set the variable outside of the distro config, however one
can refer to the value of an arbitrary variable from the local configs global
scope.