[RDSBUG-558] reintroduce /etc/version
For backwards compatibility (since internal tools might still
depend on this) the firmware version is now again also saved to
/etc/version (in addition to the new location in /etc/os-release). This
is not ideal, as /etc/version originally is a file generated deep
within the yocto build process (which is why the previous versioning
broke in the first place during a yocto update). As such, we cannot
guarantee, that this won't break again in the future. We therefore
recommend using the FIRMWARE_VERSION environment variable from
/etc/os-release instead.