Add timestamp to FIRMWARE_VERSION on dev builds
If IRMA6_DISTRO_VERSION is appended with the "-dev" suffix, any build is
treated as a non-release build and as such the PV is not sufficiant for an
exact build identification, which was an issue in the past.
To mitigate this issue, the "-dev" suffix now appends the timestamp of
the build to the FIRMWARE_VERSION variable.