gcc: Backport bugfix for bug 101510
Bug 101510 in gcc caused std::filesystem::create_directory to raise
errors on an existing symlink to a directory. This caused issues during
our unit tests within a QEMU environment.

A bugfix has been released in upstream for gcc 9.5, 10.4 and 11.3 and
newer. This commit backports this bugfix for gcc 9.3.

For more details on the upstream bug please visit
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=101510