sw-desc: Disable bootloader transaction marker
recovery_status: this is set when SWUpdate starts to write to the device, and
it is unset after the installation completed with success or it is set to
`failed` in case of error. A bootloader can use it to check if an update was
interrupted. This is a must in case a single copy of the software is on the
device. For A/B Updates it is not mandatory.