initscripts: Fix r/w access on netfitboot
Remove checkroot.sh symlink!

In /etc/fstab, the "defaults" option on the /dev/root line is replaced with
"ro" [1]. This in turn causes the init script checkroot.sh to perform a
read-only remount on startup [2]. We don't use checkroot.sh yet so
disabling the init symlink seems to be ok.

[1] https://gitlab.com/yoctoproject/poky/-/blob/dunfell/meta/classes/rootfs-postcommands.bbclass#L95
[2] https://gitlab.com/yoctoproject/poky/-/blob/dunfell/meta/recipes-core/initscripts/initscripts-1.0/checkroot.sh#L142