dropbear: Add pre generate default key for R2
The default pregenerate hostkey is only used for special cases during
development, if /etc/dropbear is not bind mounted e.g. nfsboot. Otherwise
the bind mounted host key from /mnt/iris/dropbear/dropbear_rsa_host_key
is used.

With a read-only file system, the default key path is set to /var/lib/dropbear
[1] in the rootfs if no pre generate key is available in /etc/dropbear.
That's a problem, because the mounting and bind mounting from /etc/fstab
is processed earlier:

/etc/rcS/S03mountall.sh (/etc/fstab)
/etc/rcS/S29read-only-rootfs-hook.sh (mount /var/lib/dropbear)

As a result, no bind mount can be used either to /etc/dropbear or to
/var/lib/dropbear from /mnt/iris/dropbear in /etc/fstab. With the help of the
pre generate host key under /etc/dropbear. It is not replaced [1] with
/var/lib/dropbear and the bind mount works.

[1] https://gitlab.com/yoctoproject/poky/-/blob/dunfell/meta/classes/rootfs-postcommands.bbclass#L119