dropbear: Fix potential security risk
This removes hardcoding the dropbear host key, as this poses a security
risk.

Developers flashing firmware builds in quick succession during
development and therefore experience issues with changing hostkey
fingerprints should therefore opt to connect with StrictHostKeyChecking
disabled:

```
ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null \
<sensor_ip>
```

This configuration can also be made persistent for the sensor
maintenance IP by adding the following in /etc/ssh/ssh_config:

```
Host <sensor_maintenance_ip>
    StrictHostKeyChecking no
    UserKnownHostsFile /dev/null
```

Background:

Currently, dropbear is only deployed and used during development. As
such, this is not an immidiate issue, however this distinction is not
technically enforced.

The decision that dropbear is only used internally is not set
in stone, e.g. we might decide to enable ssh access for deployment
firmware later on. In this case, the previous applied and non-obvious
modifications to dropbear would still apply, allowing an attacker
to perform a man-in-the-middle attack, as the host key is compromised.