[APC-7581] BUG: post-update command is no longer called from mongoose interface
The post-update command is no longer called when running an update via the
integrated mongoose webserver. This "bug" was probably introduced by [1].

The mongoose webserver layer has its own state machine, which is fed by the
swupdate notifications sent to it. When SWUpdate sends a message starting the
update was successful, the ipc_postupdate() function [2] is triggered. But
somehow, the "source" was no longer set to SOURCE_WEBSERVER, so this line is
skipped. The post update command will not run.

In my tests, the bug only occurs if the swu file is greater than 100kB, what causes [3] to be called.

[1] https://github.com/sbabic/swupdate/commit/d116870e943de2af088e09ccb8cce71aac7be2eb
[2] https://github.com/sbabic/swupdate/blob/master/mongoose/mongoose_interface.c#L492
[3] https://github.com/sbabic/swupdate/blob/master/mongoose/mongoose_interface.c#L640

Reported-by: Erik Schumacher <erik.schumacher@iris-sensing.com>
Reported-by: Michael Glembotzki <Michael.Glembotzki@iris-sensing.com>

Status: https://groups.google.com/g/swupdate/c/AywmYTfwe-s