[APC-7935] imx-boot: Remove invalid offset from entry point
The IVT header at the end of the fitImage is evaluated by the cst-signer
tool to generate the CSF file for signing with the Code Signing Tool
(CST). The wrong offset causes the first 0x1000 bytes to be NOT signed.

Extract from the broken CSF file created by the cst-signer tool:
Blocks = 0x48001000 0x1000 0x0164C020 "irma-fitimage.itb-ivt"

The load address and offset are calculated from the header: [1]

Presumably, in an earlier version of meta-secure-imx, the intention was to
put the IVT header at the beginning of the fitImage and the offset was
forgotten to be removed. [2] ¯\_(ツ)_/¯

[1] https://github.com/nxp-imx-support/nxp-cst-signer/blob/master/src/cst_signer.c#L1161
[2] https://source.denx.de/denx/meta-secure-imx/-/blob/dunfell/classes/fitimage-hab-sign.bbclass?ref_type=heads#L36

Signed-off-by: Michael Glembotzki <Michael.Glembotzki@iris-sensing.com>