[APC-8536] SWUpdate: Fix /swupdate/restart
With the update-types the postupdatecmd "reboot"  was moved to the "OS"
section. For performing a restart via the api call /swupdate/restart the
default postupdatecmd is called, which is empty, if not set.

This change uses reboot also as default postupdatecmd if update-type is
NOT set. The service update is untouched and still  performs no auto reboot.